+++ src/isThenable.ts
@@ -2,7 +2,7 @@ const isThenable = (val: unknown): val is Promise<unknown> => {
     if (!val) {
         return false;
     }
-    return typeof (val as {then: unknown}).then === "function";
+    return typeof (val as { then: unknown }).then === "function";
 };
 
 export {