+++ CHANGELOG.md
@@ -4,7 +4,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [UNRELEASED]
-- nothing yet
+### Dev
+- deps bump
+- tests code cleanup
+### Fixed
+- assigning undefined as data shouldn't happen
 
 ## [3.0.2] - 2021-07-06
 ### Fixed